extract KoConstructorDefinedProvider
