+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koproperty/KoPropertyDeclarationTest.kt
@@ -17,30 +17,6 @@ class KoPropertyDeclarationTest {
         sut.toString() shouldBeEqualTo sut.locationWithText
     }
 
-    @Test
-    fun `property-is-constructor-defined`() {
-        // given
-        val sut = getSnippetFile("property-is-constructor-defined")
-            .classes()
-            .properties()
-            .first()
-
-        // then
-        sut.isConstructorDefined shouldBeEqualTo true
-    }
-
-    @Test
-    fun `property-is-not-constructor-defined`() {
-        // given
-        val sut = getSnippetFile("property-is-not-constructor-defined")
-            .classes()
-            .properties()
-            .first()
-
-        // then
-        sut.isConstructorDefined shouldBeEqualTo false
-    }
-
     private fun getSnippetFile(fileName: String) =
         TestSnippetProvider.getSnippetKoScope("core/declaration/koproperty/snippet/forgeneral/", fileName)
 }
