+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koproperty/snippet/forkoconstructordefinedprovider/property-is-constructor-defined.kttxt
