+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koproperty/snippet/forkoconstructordefinedprovider/property-is-not-constructor-defined.kttxt
