+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/declaration/KoPropertyDeclarationCore.kt
@@ -3,11 +3,13 @@ package com.lemonappdev.konsist.core.declaration
 import com.intellij.psi.PsiElement
 import com.lemonappdev.konsist.api.declaration.KoKDocDeclaration
 import com.lemonappdev.konsist.api.declaration.KoPropertyDeclaration
+import com.lemonappdev.konsist.api.provider.KoConstructorDefinedProvider
 import com.lemonappdev.konsist.api.provider.KoContainingDeclarationProvider
 import com.lemonappdev.konsist.api.provider.KoKDocProvider
 import com.lemonappdev.konsist.core.cache.KoDeclarationCache
 import com.lemonappdev.konsist.core.provider.KoAnnotationProviderCore
 import com.lemonappdev.konsist.core.provider.KoBaseProviderCore
+import com.lemonappdev.konsist.core.provider.KoConstructorDefinedProviderCore
 import com.lemonappdev.konsist.core.provider.KoContainingDeclarationProviderCore
 import com.lemonappdev.konsist.core.provider.KoContainingFileProviderCore
 import com.lemonappdev.konsist.core.provider.KoDeclarationFullyQualifiedNameProviderCore
@@ -62,6 +64,7 @@ internal class KoPropertyDeclarationCore private constructor(
     KoPropertyDeclaration,
     KoBaseProviderCore,
     KoAnnotationProviderCore,
+    KoConstructorDefinedProviderCore,
     KoContainingFileProviderCore,
     KoDeclarationFullyQualifiedNameProviderCore,
     KoDelegateProviderCore,
@@ -136,8 +139,6 @@ internal class KoPropertyDeclarationCore private constructor(
             super<KoKDocProviderCore>.kDoc
         }
     }
-
-    override val isConstructorDefined: Boolean by lazy { ktCallableDeclaration is KtParameter }
 
     override fun toString(): String {
         return locationWithText
